chore: remove unused SearchableGroupEditor component
