--- src/components/SearchableGroupEditor/SearchableGroupEditor.module.css
@@ -1,4 +1,0 @@
-.header {
-    margin: 12px 0;
-    font-weight: normal;
-}
