--- src/components/SearchableGroupEditor/index.js
@@ -1,2 +1,0 @@
-import SearchableGroupEditor from './SearchableGroupEditor.js'
-export default SearchableGroupEditor
