+++ src/components/SearchableOrgUnitTree/AsyncAutoComplete/getValidationText.js
@@ -1,6 +1,23 @@
 import i18n from '@dhis2/d2-i18n'
-import createHumanErrorMessage from '../../../utils/createHumanErrorMessage'
 import { MIN_CHAR_LENGTH } from './constants.js'
+
+const createHumanErrorMessage = (
+    { message, messages, httpStatusCode },
+    fallbackMsg
+) => {
+    const fallback =
+        fallbackMsg ||
+        i18n.t('Something went wrong when processing your request.')
+    const useMessage =
+        (httpStatusCode && httpStatusCode >= 400 && httpStatusCode < 500) ||
+        (!httpStatusCode && messages && messages.length > 0)
+
+    if (!message && messages && messages.length > 0) {
+        message = messages.map(({ message }) => message).join(', ')
+    }
+
+    return useMessage ? message : fallback
+}
 
 const getValidationText = ({
     error,
